rename events files
