+++ src/custom-elements/indicator.vue
@@ -42,7 +42,7 @@
 </template>
 
 <script>
-import { getPackageInfo } from './webapp-to-content';
+import { getPackageInfo } from './webapp-events';
 import npm_logo from './assets/npm_logo.svg?component';
 import store from './store';
